Refactored redis to be a module
About to make a pubsub version
