+++ pipeline_prototype/redis/__init__.py
@@ -1,0 +1,1 @@
+"""Operations related to interacting with Redis server"""
